Update patch set 1
Patch Set 1:

I had seen that.  Which is why i wanted you to see this too.  When you get your model working in 1.x we may merge it into here and prevent much of this.  I wont merge this change because of it, just wanted to share what was needing to be done.

The only unknown is logical_device work.  In the new core_proxy there are no calls to get/create logical anything.

Patch-set: 1
Label: Code-Review=0