Update patch set 1
Patch Set 1:

Hi Lin,

Can you please remove the tabs and re-submit the code... there is a lot of diff showing up because of the whitespaces... tried configuring my settings to ignore but was not successful.

What is the approach for the voltha golang code? To use tabs or not to use tabs?

Patch-set: 1
Reviewer: Gerrit User 1000097 <1000097@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>
Label: Verified=0