Sample implementation of openolt adapter with dep for dependency management Also updated the build system to take this into account.
Note that I changed the extension of the resource manager file to hidden.
This is temporary, as the code it needs in voltha-go does not exist, and this breaks dep.
This can be worked around in development using the LOCAL_VOLTHAGO variable described in the readme
This does not build currrently, but that is due to the code in voltha-go master is missing code the openolt adapter is looking for, but I see you guys are working on it.

This pattern is consistent with how voltha-go does things, but does not leave you dependent on it to build.

See the readme for how to use dep.

Change-Id: I25b8472dbc517b193970597c9f43ddff18c2d89f