Update patch set 1
Patch Set 1:

I agree that both directions should all be binary.  We just need to be careful changing openolt agent as the voltha 1.x still depends on it behaving the "hexified" way.  If openolt agent master branch has this change then 1.x master and 2.x master will have to change to match it.  I suspect the move to BAL 3.0 might be a good time to change agent behavior as 1.7 and 2.0 will have already been tagged along with the agent tag that works with them.

Patch-set: 1
Label: Verified=0