Update patch set 2
Patch Set 2:

part of the problem is jenkins runs a "dep check" on commit and thats failing because vendored voltha-go filed were manually copied into the vendor folder and checked in.   Setting LOCAL_VOLTHAGO does this to make working on both voltha-go and voltha-openolt-adapter easier, but it cannot be checked in that way.

What will have to happen is the voltha-go change will have to be committed first, then in voltha-openolt-adapter run a "dep ensure; dep ensure -update github.com/opencord/voltha-go" which will update the vendored items the dep way and bump the Gopkg.lock file appropriately.

Patch-set: 2