Update patch set 3
Patch Set 3:

Yes David, I think that these changes are needed for the 2.1 branch as well. Can I do the cherry pick to that branch?

What would be the general process, is it the author who is responsible for the cherry pick to the different branches?

Patch-set: 3
Reviewer: Gerrit User 1000097 <1000097@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>
Label: Code-Review=0