Update patch set 1
Patch Set 1: Code-Review-1

(3 comments)

In general, I am against the thought that cache/stateful-data is an issue for reconciliation. Cache improves performance, and as long as we can re-build the cache on voltha restarts we should do that. Frequent KV access decreases performance, esp. if we keep doing that for every flow/packet-in/packet-out.
I could not spend enough time analyzing each case, but if you think certain information, like for ex: NniIntfId does not change during the lifetime of the OLT, then such information should be cached. Cache the information while PUT to KV store, and then return from the cache during GET. The cache could be invalidated during voltha-restart/olt-reboot etc. and then could be re-build again. 
So I request you to check case-by-case and use cache where necessary. Back-up on KV store, but cache locally too if possible and if that makes sense.

Patch-set: 1
Label: Code-Review=-1