Update patch set 4
Patch Set 4:

The changes in this review are replicating the flows across pbits/gem-ids and this is exhausting trap flows on the BAL. I had raised ticket on Broadcom CS9196966 and Broadcom said they are not going to fix this issue in BAL3.1 and BAL3.2 has a different mechanism using access_control object to trap a class of packets. More details about access_control object on BAL3.2_EA release notes. So, for now I suggest to abandon this review and re-work a suitable solution later. Also, 1T4G works anyway as of today, with the limitation to use pbit0 only (not really great, but it is what it is).

Patch-set: 4
Reviewer: Gerrit User 1000127 <1000127@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>
Label: Code-Review=0