Update patch set 7
Patch Set 7: Code-Review-1

(3 comments)

Please revisit the changes per the comments. Review the entire file again, as I could not go line by for where it needs lock(or Rlock) and Unlock (or RUnlock). Run the test tests that causes the issue reported in VOL-2108 to verify that it is indeed fixed and then it should be ok.

Patch-set: 7
Label: Code-Review=-1