Update patch set 1
Patch Set 1:

Without following the code, if I were just to go by the patch description, it says "during flow removal which failed to remove default meter when all flows are deleted for a subscriber"

It is perfectly ok to have all flows removed for a meter and yet not have the meter be deleted (default or otherwise). It is the job of ONOS to actually send a meter-delete command for a meter that is no longer needed. VOLTHA should not automatically (on its own) delete things for state managed by ONOS (such as flows/groups/meters etc). It needs to be explicitly told to do so by ONOS 

So my question is
1) is this patch really automatically deleting things 
OR
2) is the description incorrect?

Patch-set: 1
Label: Verified=0