Update patch set 8
Patch Set 8:

I think this change and https://gerrit.opencord.org/#/c/16218/ may be deadlocked given they both try to move to voltha-lib-go 2.2.15 but other changes must be made to openolt adapter before it can be used.

Those changes Girish seems to be making in https://gerrit.opencord.org/#/c/16218/ but it seems some of the mock changes in this patch are needed to pass lint.

Can we try to set this patch to voltha-lib-go 2.2.14 and see if we can get it to pass with that?  Then the other patch can move to 2.2.15?

Patch-set: 8