Update patch set 10
Patch Set 10:

The https://jenkins.opencord.org/job/voltha-openolt-adapter_sanity-system-test/274/ has failed by the way. It is the hardware test that has passed. I am not sure if this is good to merge yet. 
The voltha-openolt-adapter_sanity-system-test complains about failure to delete the device. It is used to pass earlier, not sure what the problem is. I have also reported this on #voltha-stabilization slack channel.

Patch-set: 10