Update patch set 8
Patch Set 8:

> (10 comments)

There seems to be a lot of inconsistency with the log messages in openolt adapter. At some places it is dasherized and some places it is not. I always liked it be consistent with the way other voltha projects are doing it (voltha-go, voltha-openonu-adapter, pyvoltha projects has all logs dasherized). So, we dasherize or not? I am ok either way, I just need the logging to be consistent across all voltha projects.

Patch-set: 8
Label: Code-Review=0