Update patch set 30
Patch Set 30:

There are places in the code where error handling was added, which is good, but the handling in simply logging the error. Again, this is better than what we currently have, but is there more that should be done? i.e. post the error on kafka? change behavior or system state because of the failure?

Patch-set: 30
Label: Verified=0