Update patch set 2
Patch Set 2:

> (1 comment)

I am not sure I follow what you said. The change was to update the local cache of gemports to be in sync with what is present on external KV store. When OLT Device is deleted, the external KV store is updated anyway correctly now. It does not matter we update the local cache or not in this case, because the OLT device itself is getting deleted.

Patch-set: 2