Update patch set 8
Patch Set 8:

> (1 comment)

If the ChildDeviceDetected times out, the adapter should free the ONU ID. As you said the core may have some stale data about the ONU. To clean that up, the adapter should send a ChildDeviceLost right after freeing the ONU ID and it need not wait for core to respond on ChildDeviceLost (call it in a go routine).  This way core does cleanup (if it finds any references).
But, looking up on the CoreProxy interface, I do not find any RPC for ChildDeviceLost. I think it makes sense to have one, especially for cases like this.
There is ChildDevicesLost, and it looks like it is meant for all devices on the OLT.
The ChildDeviceLost should have arguments on the same lines as ChildDeviceDetected.
Hope this makes sense.

Patch-set: 8