Update patch set 10
Patch Set 10: -Code-Review

> According to my investigation the issue is that "dh.coreProxy.ChildDeviceDetected"
 > times out but the device is created.
 > 
 > A solution could be to remove the device if this happens, but I'm
 > concerned it will result in an endless timeout loop

Deleting the device from the core following the ChildDeviceDetected is a good idea. But, as you say, the DeleteDevice could also timeout. If that happens can we store the serial number of this ONU? When we receive the Discovery for this ONU again, we should first the check the ONU serial number against previous timed out serial number list/map - if found, try deleting that device from the core first. 
If we again see a timeout, it follows the old handling, i.e., store in timed-out list/map.
This sounds messy, but this is indeed a messy situation. I am not sure if there is another way to handle this.

Patch-set: 10
Label: Code-Review=0