Update patch set 38
Patch Set 38: Code-Review-2

the func `addFlowToDevice` is called several places and the value of that func is used even if err is not nil. this is technically not correct usage and these uses need to be changed.

Patch-set: 38
Reviewer: Gerrit User 1000003 <1000003@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>
Label: Code-Review=-2