Update patch set 6
Patch Set 6:

ok, not sure things are well in this patch yet. or maybe one of the later updates took something in the wrong direction.

create the bbsim device (success)
create the same bbsim device (fails as expected)
enable the olt (goes to activating state, but connection state as unknown and never enables)

Patch-set: 6