Update patch set 10
Patch Set 10:

Sanity tests fail because BBSIM is programmed to send the DHCP discovery message only when it gets the DHCP trap flow with pbit 0. When the pbit value is set to 0b11111111 (which is default), 8 DHCP flows are pushed by the openolt adapter. This patch pushes one flow if the pbit is 0b11111111(a DHCP flow with pbit=255)

Patch-set: 10
Label: Code-Review=0