Update patch set 7
Patch Set 7:

> Dev, please review https://gerrit.opencord.org/#/c/18852/.
 > Please use the new message format (as seen my patch) to configure
 > the RDI error alarm.
 > I have documented in my earlier comment on how you need to
 > interpret the message to configure the RDI alarm based on
 > RATE_THRESHOLD.

OK so for each onu a default set of values needs to be pushed and also a cache of the threshold of rdi error count needs to be kept in the adapter for each onu and every time the adapter receives an onuITUPonStats ind the value of the rdi_errors needs to be compared to this cache to raise/clear the alarm. Is this correct?

Patch-set: 7