Update patch set 4
Patch Set 4:

> > (1 comment)
 > 
 > In my tests with BAL3.4 integrated agent I receive
 > BCMOLT_ONU_AUTO_SUBGROUP_ONU_DEACTIVATION_COMPLETED on onu soft
 > reboot. The oper_state and admin_state are both sent as "down" from
 > the agent following this.
 > 
 > Below are also some relevant logs from the agent fyi.
 > 
 > [3253553: I OPENOLT             ] indications.cc 204| LOS
 > indication : intf_id: 0 port: 536870912 status on
 > [3253803: I OPENOLT             ] indications.cc 879| Got onu
 > deactivation, intf_id 0, onu_id 1, fail_reason 0, result_status
 > success
 > [3253803: I OPENOLT             ] indications.cc 895| received onu
 > deactivate result, pon intf 0, onu_id 1, status 0, reason 0

If i understand correctly, in case of ONU soft reboot, the BAL is deactivating the ONUs and sending a deactivation complete indication with admin state and operation state set to UP.
As per my knowledge, in this scenario, the BAL should not deactivate the ONUs and should send ONU indication with operState set to Down(Still the Admin state of ONU is UP). I assume this is a bug in OLT and need to be fixed in device.

As the changes are reflected whenever ONU goes down, have you tested the scenarios where the ONU operationally goes down and remain administratively up in the device.

Patch-set: 4
Label: Code-Review=0