[VOL-3644] fixed wrong topic names
The adapter_open_olt_topic is a configurable parameter from the command line

When sending InterAdapterMessage to the openonu-adapter, openolt-adapter
uses hardcoded "openolt" or deviceType fields as source topic name. This causes
not getting responses from the This causes not getting responses from the openonu-adapter
if the topic name argument is set different than "openolt".

Change-Id: I0f88c1e8f9bfe322344c062478593444466869d4