Update patch set 4
Patch Set 4:

(2 comments)

> Patch Set 4:
> 
> (2 comments)
> 
> In general I understand the reason for the change from the jira. But can you clarify how this change worked earlier. Has something changed now, like the OLT is not sending serial number anymore - but nothing changed on OLT.
> Also is using per PON channel to queue ONU indications have any effect on performance unlike earlier with per ONU channel. Does scale test show any change in performance data (good or bad or no change).

In BBSIM case, all indications include serial number so the previous patch was working well. But in Edgecore case, the indications don't include serial number so all indications from all ONUs were queued on the same channel with serial number "". It was working and ensuring the ordering of onuIndications, but not parallel.

We didn't see any effect on performance hence this is only queueing the onuInd and onuDiscInd, not the omciInd. Still have enough threads for two indications.

Patch-set: 4