Update patch set 6
Patch Set 6:

The K8s issue was fixed, but there is an issue with the BNG emulator on the verification pod, working to fix it. Once that is fixed will retrigger the tests here again. The pod has rebooted in the last few days and things are not in the right shape after it was back.

Patch-set: 6
Reviewer: Gerrit User 1000127 <1000127@2a2bfe1b-c5c2-48ed-9ac8-16438ab24388>