VOL-4251: Delete device immediately after (without any delay) volt-remove-subscriber access leaves stale resources on the OLT device
The delete device cleans up everything for the device including the
TpInstance even before volt-remove-subscriber-access has completed
the processing. The volt-remove-subscriber cleans up scheduler
and queues at the end of removing all the flows. But it needs the
tp-instance reference to clean up the schedulers and queues, but
it has already been deleted. So, it aborts the scheduler/queue
cleanup on the OLT and this causes issues when configuring the
scheduler again on the OLT during a fresh setup as there are some
stale entries on the ONU.

The fix here is to force cleanup the resources on the OLT when the
ONU device is being deleted on the OLT.

Change-Id: I54cd3ef0d5bd41cd901f3bb8917927336b84ea27
(cherry picked from commit 197acc100ab4eab2354ac46700ece874574b3000)