Update patch set 2
Patch Set 2:

The failed hw test is the last test in the suite - DisableEnablePonPort and unrelated to this patch. One of the ONU could not complete eap auth and from the robot logs the downstream packets could not reach the ONU (while the upstream passed fine). From the past experience I have seen this happen if the ONU flows flows and tp were configured in a particular order the alpha ONU had some issues with downstream traffic, but i have not investigated if that was the problem here. Since this issue is unrelated to my patch I am going to re-trigger the hw tests.

Patch-set: 2